feat: introduce adobe core martech stack
